KMS: encrypt() should allow an alias as argument
